+++ response.js
@@ -15,8 +15,7 @@ exports = module.exports = function(response) {
 res.send = function(chunk, type) {
   if (!chunk) throw new TypeError('res.send() argument chunk is required');
   if (typeof chunk === 'string') this.setHeader('Content-Type', mime.lookup(type ? type : 'txt'));
-  if (this.cache)
-    if (cache.check(this.req, etag(chunk))) return;
+  if (this.cache) if (cache.check(this.req, etag(chunk))) return;
   this.end(Buffer.from(chunk, 'utf8'), 'utf8');
   return;
 };