+++ server.js
@@ -17,4 +17,5 @@ exports.router = require('./router.js');
 exports.cache = require('./cache.js');
 exports.cors = require('./cors.js');
 exports.cookie = require('./cookie.js');
+exports.parser = require('./parser.js');
 exports.utilities = require('./utilities.js');